refactor: Move Open API 2/3 specific response checks to a separate module
